// examples/workflow/common/src/lib/configuration/contract-confirm/enum/contract-confirm-containers.enum.ts
/*
 * Copyright(c) RIB Software GmbH
 */

/**
 * Uuids of containers available in contract confirm wizard.
 */
export enum ContractConfirmWizardContainers {
	CON_DETAIL = 'B3B0FDF482AE4973A4B6BBEA754876C3',
	CON_LIST ='E5B91A61DBDD4276B3D92DDC84470162',
	CON_BUSINESSPARTNER = '931850b088e64590a395b0eb21e5f4dc',
	CONTRACT_CONFIRM_COVER_LETTER ='55bc306e62014bb6aa03336eb58d8c51',
	CONTRACT_CONFIRM_REPORT = '8777043964bd45caab4f58af28a6a0b7',
	CONTRACT_CONFIRM_PROJECT_DOCUMENT = '9fe6a21dfcca4747b817480e059db4d0',
	CONTRACT_CONFIRM_DOCUMENT = '3473623dccd04ef7ad14a2cf18e74eb5',
	CONTRACT_PRC_BOQ = 'A56A75CBE90545ECBFAFA5DE3F437F10',
}