// examples/workflow/common/src/lib/configuration/rfq-bidder/enum/rfq-bidder-containers.enum.ts
/*
 * Copyright(c) RIB Software GmbH
 */

/**
 * Uuids of containers available in rfq bidder.
 */
export enum RfqBidderWizardContainers {
	RFQ_BIDDER = '2dd4dc4f50844b22b55f5815f83fed2e',
	RFQ_BIDDER_SETTINGS = 'e907d30b10274a1fb760f76b113c3b0d',
	RFQ_DATA_FORMAT = '9d4498f32ca046ecb961fba5bac6436d',
	RFQ_SENDER = '44e5b148b85845c5bb13e35f7c97c5c6',
	RFQ_BIDDER_DOCUMENTS_PROJECT = 'fc2ce31089fa456587193a13bcab3a43',
	RFQ_BIDDER_DOCUMENTS = '7016ed0f76da4be19b63c516d75891f6',
	RFQ_BIDDER_Report = '42ede45a94ce421f822305090bbd0915',
	RFQ_BIDDER_COVER_LETTER = '7ba78cfb8fd242ec8eb4190a8ba3559f',
	RFQ_PROCUREMENT_STRUCTURE_DOCUMENTS = '851d650c916a42778042f7730e48198a',
	RFQ_CLERKDOCUMENTS = '1e7537f6803c4bebb821634bb17cb068',
	RFQ_BIDDER_BOQ_SELECT = 'c31cc0077055459d8000a6ebbf523cb6',
	TRANSMISSION = 'b6a85fec74a2402fb3bc16f108c9c0b3'
}